requirements-test.txt
flake8
django-adv-cache-tag==1.1.0
